+++ src/components/app/tma/TmaStakeUnstake.tsx
@@ -82,6 +82,7 @@ const TmaStakeUnstake = observer(({ model }: Props) => {
             onInput={(e) => {
               model.setAmount(e.currentTarget.value)
             }}
+            onBlur={model.normalizeAmount}
             onKeyDown={(e) => {
               if (e.key === 'Enter' && model.isButtonEnabled) {
                 const button = document.querySelector<HTMLInputElement>('#submit')